调整命名空间

+++ IEC104/DatagramFormatters/C_RD_NA_1.cs
@@ -18,7 +18,7 @@ namespace Shouyuan.IEC104
 
         }
 
-        public APDU Create(byte asduaddr, byte msgAddr, byte cot = 5)
+        public APDU Create(byte asduaddr, uint msgAddr, byte cot = 5)
         {
             var apdu = CreateAPDU(asduaddr);
             apdu.ASDU.Cause = cot;
